implement builtin_echo

+++ kernel/app/hello.c
@@ -2,7 +2,7 @@
 #include <barelib.h>
 
 /* I have to put this somewhere... */
-int strcmp(const char* str1, const char* str2) {
+int16 strcmp(const char* str1, const char* str2) {
   for(; *str1 == *str2; str1++, str2++) {
     if(*str1 == '\0') { return 0; }
   }
